.SG Bus Info Chatbot/requirements.txt
Cython==0.27.3
Flask==0.10.1
json2html==1.2.1
numpy==1.14.0
opencv-python==3.4.0.12
Pillow==6.2.0
pytesseract==0.1.7
requests==2.18.4
tesseract
urllib3==1.22
watson-developer-cloud==1.0.2
Werkzeug==0.14.1
